clean repository unecessary folders and files, general reorganizations and dockstrings, new readme
