+++ src/api/routers/medication_drugs.py
@@ -2,8 +2,8 @@ from fastapi import APIRouter, HTTPException
 from sqlmodel import Session, select
 from sqlalchemy.orm import selectinload
 from db.manager import Database
-from api.schemas.models import *
-from api.schemas.schemas import *
+from db.models import *
+from api.schemas import *
 from typing import List
 
 drugs_router = APIRouter()
@@ -115,13 +115,13 @@ def deactivate_user_drug(user_id: int, drug_id: int):
             select(DrugUse)
             .where(DrugUse.id == drug_id)
             .where(DrugUse.user_id == user_id)
-            .where(DrugUse.status == "Active")
+            .where(DrugUse.status == "active")
         ).first()
 
         if not drug:
             raise HTTPException(status_code=404, detail="Drug not found")
 
-        drug.status = "Inactive"
+        drug.status = "inactive"
         drug.end_date = datetime.now().strftime("%Y-%m-%d")
 
         session.add(drug)
